+++ src/main/java/salute/oneshot/domain/ingredient/repository/IngrElasticQueryRepository.java
@@ -8,17 +8,18 @@ import co.elastic.clients.elasticsearch._types.query_dsl.QueryBuilders;
 import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.SearchResponse;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import salute.oneshot.domain.ingredient.dto.service.SearchIngrSDto;
 import salute.oneshot.domain.ingredient.entity.IngredientDocument;
 import java.io.IOException;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
-public class IngredientSearchFinder {
+public class IngrElasticQueryRepository {
 
     private final ElasticsearchClient client;
-
     private final String INGREDIENT_INDEX = "ingredients";
 
     public SearchResponse<IngredientDocument> ingrSearchByCondition(SearchIngrSDto sDto) throws IOException {
@@ -29,14 +30,19 @@ public class IngredientSearchFinder {
 
         BoolQuery.Builder builder = QueryBuilders.bool();
 
-        if (sDto.getCategory() != null) {
-            builder.filter(Query.of(q -> q.term(m -> m.field("category").value(sDto.getCategory()))));
+        if (!sDto.getCategory().isBlank()) {
+            builder.filter(Query.of(p -> p.term(m -> m.field("category").value(sDto.getCategory()))));
         }
 
         if (!sDto.getKeyword().isBlank()) {
-            addShouldIfNotNull(builder, sDto.getKeyword(), "name", 3.0f);
-            addShouldIfNotNull(builder, sDto.getKeyword(), "description", 2.0f);
+
+            addShouldIfNotNull(builder, sDto.getKeyword(), "name", 2.0f);
+            addShouldIfNotNull(builder, sDto.getKeyword(), "description", 1.0f);
+
         }
+
+        builder.minimumShouldMatch("1");
+
 
         SearchRequest searchRequest = new SearchRequest.Builder()
                 .index(INGREDIENT_INDEX)
@@ -50,7 +56,9 @@ public class IngredientSearchFinder {
     }
 
     private void addShouldIfNotNull(BoolQuery.Builder builder, String condition, String fieldName, float boost) {
-        builder.should(Query.of(q -> q.match(m -> m.field(fieldName)
-                .query(condition.toLowerCase()).boost(boost))));
+        builder.should(Query.of(q -> q.matchPhrase(m -> m.field(fieldName)
+                .query(condition).boost(boost))));
     }
 }
+
+
